Convert CLI top-level imports as dynamic imports (#7635)
